# internal/app/openapi.yaml
openapi: 3.1.0
info:
  title: WaCalls API
  version: 0.1.0
  license:
    name: MIT
    identifier: MIT
  description: >-
    HTTP API for WaCalls, a self-hosted WhatsApp calling bridge. All /api/*
    endpoints require the bearer token configured via WACALLS_API_TOKEN
    (when unset, authentication is disabled). The SSE stream also accepts the
    token as the access_token query parameter. This document is the canonical
    contract: a test gate fails the build when routes and spec drift apart.
servers:
  - url: /
security:
  - bearerAuth: []
  - queryToken: []
paths:
...
  /api/sessions:
    get:
      operationId: listSessions
      summary: List WhatsApp accounts
      responses:
        "200":
          description: All configured sessions
          content:
            application/json:
              schema:
                type: object
                required: [sessions]
                properties:
                  sessions:
                    type: array
                    items:
                      $ref: "#/components/schemas/SessionInfo"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "429": { $ref: "#/components/responses/RateLimited" }
    post:
      operationId: createSession
      summary: Create an account and begin QR pairing
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: Display name, defaults to "Session"
      responses:
        "200":
          description: Session created
          content:
            application/json:
              schema:
                type: object
                required: [id]
                properties:
                  id:
                    type: string
        "401": { $ref: "#/components/responses/Unauthorized" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "500": { $ref: "#/components/responses/InternalError" }
  /api/sessions/{sid}:
    delete:
      operationId: deleteSession
      summary: Log out and remove an account
      parameters:
        - $ref: "#/components/parameters/Sid"
      responses:
        "204":
          description: Session removed
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /api/sessions/{sid}/logout:
    post:
      operationId: logoutSession
      summary: Disconnect an account, keeping it for re-pairing
      parameters:
        - $ref: "#/components/parameters/Sid"
      responses:
        "204":
          description: Session disconnected
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /api/sessions/{sid}/pair:
    post:
      operationId: pairSession
      summary: Re-pair an account, emitting a fresh QR over SSE
      parameters:
        - $ref: "#/components/parameters/Sid"
      responses:
        "204":
          description: Pairing restarted
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /api/sessions/{sid}/pair-code:
    post:
      operationId: pairSessionByPhoneCode
      summary: Re-pair an account with an 8-digit code instead of a QR
      description: >
        Starts the "Link with phone number instead" flow: a fresh device is created,
        connected, and WhatsApp is asked for a pairing code that the operator types
        into the phone. Use this when the phone cannot scan the browser's screen.
        The code is also pushed over SSE (session-auth-state, state="pair_code") and
        appears in SessionInfo.code until the link succeeds or times out.
      parameters:
        - $ref: "#/components/parameters/Sid"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone]
              properties:
                phone:
                  type: string
                  description: >
                    The phone being linked, international format with the country
                    code first. Non-digits are stripped; at least 8 digits required.
                  example: "5511987654321"
      responses:
        "200":
          description: Pairing code issued
          content:
            application/json:
              schema:
                type: object
                required: [code]
                properties:
                  code:
                    type: string
                    description: 8-digit code to enter on the phone
                    example: "ABCD1234"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /api/sessions/{sid}/presence:
    post:
      operationId: setSessionPresence
      summary: Send presence ping or update status for an account
      description: >
        Sends an explicit presence update to WhatsApp for this session. Defaults to
        "available" if no body or state is specified. Used for anti-inactivity
        keepalive and external presence synchronization.
      parameters:
        - $ref: "#/components/parameters/Sid"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                state:
                  type: string
                  description: Presence state ('available' or 'unavailable')
                  default: "available"
                  example: "available"
      responses:
        "200":
          description: Presence update sent
          content:
            application/json:
              schema:
                type: object
                required: [status, session, presence]
                properties:
                  status:
                    type: string
                    example: "ok"
                  session:
                    type: string
                  presence:
                    type: string
                    example: "available"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "502": { $ref: "#/components/responses/InternalError" }
  /api/sessions/{sid}/qr:
    get:
      operationId: getSessionQR
      summary: Current pairing QR payload of the account
      description: >
        Returns the QR string last emitted for this session, or an empty string
        once the account is paired. The same payload is pushed over SSE as
        session-qr; poll this only when the stream is unavailable.
      parameters:
        - $ref: "#/components/parameters/Sid"
      responses:
        "200":
          description: QR payload (empty when already paired)
          content:
            application/json:
              schema:
                type: object
                required: [qr]
                properties:
                  qr:
                    type: string
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /api/sessions/{sid}/status:
    get:
      operationId: getSessionStatus
      summary: Connection state of one account
      parameters:
        - $ref: "#/components/parameters/Sid"
      responses:
        "200":
          description: Session state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionInfo"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /api/sessions/{sid}/calls:
    post:
      operationId: startCall
      summary: Start an outgoing call
      parameters:
        - $ref: "#/components/parameters/Sid"
        - $ref: "#/components/parameters/ClientId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone]
              properties:
                phone:
                  type: string
                  description: E.164-ish phone number; non-digits are stripped
      responses:
        "200": { $ref: "#/components/responses/CallRef" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: Operator identified by X-Client-Id is already on a call
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          description: Per-session concurrent call cap reached, or rate limited
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500": { $ref: "#/components/responses/InternalError" }
        "503":
          description: Session is not paired
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    get:
      operationId: listCalls
      summary: Live calls of the session, same shape as the SSE call-list
      parameters:
        - $ref: "#/components/parameters/Sid"
      responses:
        "200":
          description: Live calls ordered by startedAt then callId
          content:
            application/json:
              schema:
                type: object
                required: [calls]
                properties:
                  calls:
                    type: array
                    items:
                      $ref: "#/components/schemas/CallRecord"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /api/sessions/{sid}/calls/{id}:
    get:
      operationId: getCall
      summary: One live call
      parameters:
        - $ref: "#/components/parameters/Sid"
        - $ref: "#/components/parameters/CallId"
      responses:
        "200":
          description: The live call; 404 once it ends or if it belongs to another session
          content:
            application/json:
              schema:
                type: object
                required: [call]
                properties:
                  call:
                    $ref: "#/components/schemas/CallRecord"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { $ref: "#/components/responses/RateLimited" }
    delete:
      operationId: endCall
      summary: End an active call
      parameters:
        - $ref: "#/components/parameters/Sid"
        - $ref: "#/components/parameters/CallId"
      responses:
        "204":
          description: Call ended (idempotent)
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /api/sessions/{sid}/calls/{id}/webrtc:
    post:
      operationId: exchangeWebRTC
      summary: Exchange the browser WebRTC SDP for the call audio bridge
      parameters:
        - $ref: "#/components/parameters/Sid"
        - $ref: "#/components/parameters/CallId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [sdp_offer]
              properties:
                sdp_offer:
                  type: string
      responses:
        "200":
          description: SDP answer
          content:
            application/json:
              schema:
                type: object
                required: [sdp_answer]
                properties:
                  sdp_answer:
                    type: string
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "500": { $ref: "#/components/responses/InternalError" }
  /api/sessions/{sid}/calls/{id}/ws:
    get:
      operationId: attachCallWebSocket
      summary: Attach the operator's audio to a call over a WebSocket (WebRTC fallback)
      description: >
        Upgrades to a WebSocket that carries the operator leg of the call, for
        browsers that cannot reach the media port over UDP (a firewall dropping UDP,
        or a reverse proxy with no UDP path). It replaces whatever transport was
        attached to the call before, so it also works for picking up a transferred
        call.


        Wire format, both directions: one binary message per frame of raw Int16
        little-endian PCM, 16 kHz mono — exactly what the WebRTC "pcm" data channel
        carries. Text messages and frames larger than 16 KiB are ignored. The client
        MUST request the "pcm16" subprotocol. Audio only: a video call still needs
        the WebRTC transport. The server pings every 20 s to survive idle-connection
        reapers; when the socket drops, the call ends.


        Authentication is the same as every other /api route, but a cross-origin
        handshake cannot carry the SameSite=Strict session cookie, so an operator on
        another origin must pass ?apiKey= (or ?access_token=) on the query string.
      parameters:
        - $ref: "#/components/parameters/Sid"
        - $ref: "#/components/parameters/CallId"
      responses:
        "101":
          description: Switching protocols; the socket now carries PCM frames
        "400":
          description: Not a valid WebSocket upgrade, or the Origin is not allowed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /api/sessions/{sid}/calls/{id}/accept:
    post:
      operationId: acceptCall
      summary: Accept an incoming call, claiming it for the X-Client-Id operator
      parameters:
        - $ref: "#/components/parameters/Sid"
        - $ref: "#/components/parameters/CallId"
        - $ref: "#/components/parameters/ClientId"
      responses:
        "200": { $ref: "#/components/responses/CallRef" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: Claimed by another client, or operator already on a call
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429": { $ref: "#/components/responses/RateLimited" }
        "500": { $ref: "#/components/responses/InternalError" }
  /api/sessions/{sid}/calls/{id}/reject:
    post:
      operationId: rejectCall
      summary: Reject an incoming call
      parameters:
        - $ref: "#/components/parameters/Sid"
        - $ref: "#/components/parameters/CallId"
      responses:
        "200":
          description: Rejected (idempotent when the call is already gone)
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties:
                  status:
                    type: string
                    const: ok
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: Call is in a state that cannot be rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429": { $ref: "#/components/responses/RateLimited" }
  /api/sessions/{sid}/calls/{id}/video:
    post:
      operationId: enableVideo
      summary: Turn the local camera on or off mid-call
      description: >
        With no body (or any action other than "stop"), sends a WhatsApp mid-call
        video-upgrade request (<video state=3>) to the peer and derives the video
        SSRCs so the relay forwards both video streams; the browser starts
        capturing/encoding once this returns. With {"action":"stop"}, signals
        <video state=6> so the peer's UI drops our tile instead of freezing on the
        last decoded frame — only the outbound direction stops, inbound video and the
        relay subscription are left alone so a later enable needs no renegotiation.
      parameters:
        - $ref: "#/components/parameters/Sid"
        - $ref: "#/components/parameters/CallId"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                action:
                  type: string
                  enum: [enable, stop]
                  default: enable
                  description: Anything other than "stop" requests the video upgrade
      responses:
        "200":
          description: Video state change signalled
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties:
                  status:
                    type: string
                    const: ok
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "500": { $ref: "#/components/responses/InternalError" }
  /api/sessions/{sid}/calls/{id}/hold:
    post:
      operationId: holdCall
      summary: Park the WhatsApp peer on hold music, or resume the call
      description: >
        On hold, the operator's microphone stops being relayed and hold music is
        injected into the WhatsApp leg instead. The music is read per hold from
        DataDir/hold-music.wav when present (uploaded via POST /api/holdmusic),
        otherwise the built-in chime is used. Broadcasts call-hold-state over SSE
        so every operator's UI shows the badge.
      parameters:
        - $ref: "#/components/parameters/Sid"
        - $ref: "#/components/parameters/CallId"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                action:
                  type: string
                  enum: [hold, resume]
                  default: hold
                  description: Anything other than "resume" holds the call
      responses:
        "200":
          description: Hold state applied
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties:
                  status:
                    type: string
                    const: ok
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "500": { $ref: "#/components/responses/InternalError" }
  /api/sessions/{sid}/calls/{id}/transfer:
    post:
      operationId: transferCall
      summary: Blind-transfer a live call to another operator
      description: >
        Transfers between operators on the SAME paired WhatsApp account: the
        WhatsApp leg (CallManager) stays up, the call goes on hold so the peer
        hears music, and a call-transfer SSE event prompts the target operator.
        When that operator attaches its browser via POST .../webrtc, the new
        bridge replaces the old one and the call resumes automatically.
      parameters:
        - $ref: "#/components/parameters/Sid"
        - $ref: "#/components/parameters/CallId"
        - $ref: "#/components/parameters/ClientId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [to]
              properties:
                to:
                  type: string
                  description: Target operator identity (the X-Client-Id it connects with)
      responses:
        "200":
          description: Transfer offered to the target operator
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties:
                  status:
                    type: string
                    const: ok
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "500": { $ref: "#/components/responses/InternalError" }
  /api/sessions/{sid}/calls/{id}/record:
    post:
      operationId: recordCall
      summary: Start or stop server-side recording of a call
      description: >
        Mixes both legs (peer audio and operator microphone) into a single mono
        16 kHz 16-bit PCM WAV at RecordingsDir/<callID>.wav. Starting twice is a
        no-op; the recording is closed automatically when the call ends.
        Broadcasts call-record-state over SSE.
      parameters:
        - $ref: "#/components/parameters/Sid"
        - $ref: "#/components/parameters/CallId"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                action:
                  type: string
                  enum: [start, stop]
                  default: start
                  description: Anything other than "stop" starts recording
      responses:
        "200":
          description: Recording state applied
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties:
                  status:
                    type: string
                    const: ok
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "500": { $ref: "#/components/responses/InternalError" }
  /api/holdmusic:
    get:
      operationId: getHoldMusic
      summary: Status of the hold-music file
      description: >
        Reports whether an admin-uploaded WAV is present in DataDir. When absent,
        exists is false and name describes the built-in chime.
      responses:
        "200":
          description: Hold-music status
          content:
            application/json:
              schema:
                type: object
                required: [exists, name, size]
                properties:
                  exists:
                    type: boolean
                  name:
                    type: string
                  size:
                    type: integer
                    description: Size in bytes, 0 when using the built-in chime
        "401": { $ref: "#/components/responses/Unauthorized" }
        "429": { $ref: "#/components/responses/RateLimited" }
    post:
      operationId: uploadHoldMusic
      summary: Replace the hold-music file
      description: >
        Uploads the WAV played to the peer while a call is on hold. The upload must
        be a 16-bit PCM RIFF/WAVE file (mono or stereo, 8-48 kHz); anything else is
        rejected so a bad upload cannot silently fall back to the built-in chime.
        The server downmixes and resamples it to 16 kHz mono — the rate the
        WhatsApp leg runs at — and writes the normalised result atomically to
        DataDir/hold-music.wav. Unlike the rest of /api/*, this route allows bodies
        above the 1 MiB cap (limit is 8 MiB).
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: 16-bit PCM WAV, mono or stereo, 8-48 kHz
      responses:
        "200":
          description: Hold music replaced
          content:
            application/json:
              schema:
                type: object
                required: [status, size]
                properties:
                  status:
                    type: string
                    const: ok
                  size:
                    type: integer
                    description: Size in bytes of the stored 16 kHz mono file
                  source_rate:
                    type: integer
                    description: Sample rate of the uploaded file, before conversion
                  source_chans:
                    type: integer
                    description: Channel count of the uploaded file, before downmixing
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "413":
          description: Upload exceeds the 8 MiB limit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429": { $ref: "#/components/responses/RateLimited" }
        "500": { $ref: "#/components/responses/InternalError" }
...
  /api/recordings/{id}:
    get:
      operationId: downloadRecording
      summary: Download one call recording
      description: >
        Streams the 16 kHz mono WAV for a call, with Range support so a player can
        seek. When the call is still live the response carries X-Recording-Active:
        true and the audio will read as empty, because the length field in the WAV
        header is only patched when recording stops.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
          description: Call ID, as returned by /api/recordings
      responses:
        "200":
          description: The recording
          headers:
            X-Recording-Active:
              description: Present and true while the call is still being recorded
              schema: { type: string }
          content:
            audio/wav:
              schema:
                type: string
                format: binary
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "500": { $ref: "#/components/responses/InternalError" }
...
  /api/logout:
    post:
      operationId: logout
      summary: Invalidate the current session and clear the cookie
      responses:
        "204":
          description: Logged out
        "401": { $ref: "#/components/responses/Unauthorized" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    queryToken:
      type: apiKey
      in: query
      name: access_token
  parameters:
    Sid:
      name: sid
      in: path
      required: true
      schema:
        type: string
      description: Session id
...
  responses:
    Unauthorized:
      description: Missing or invalid token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    RateLimited:
      description: Per-IP rate limit exceeded
      headers:
        Retry-After:
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    InternalError:
      description: Unexpected failure
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
    SessionInfo:
      type: object
      required: [id, name, jid, state, paired]
      properties:
        id:
          type: string
        name:
          type: string
        jid:
          type: string
        state:
          type: string
        paired:
          type: boolean
        qr:
          type: string
          description: Pairing QR payload; empty once paired
        code:
          type: string
          description: >
            8-digit pairing code, present only while a POST /pair-code flow is in
            flight. Mutually exclusive with qr.
...